[feat]리팩토링 완성
