+++ app/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".feature.MainActivity"
+            android:name=".feature.main.MainActivity"
             android:exported="false"
             android:label="@string/app_name"
             android:theme="@style/Theme.NOWSOPTAndroid"></activity>
