+++ app/src/main/java/com/sopt/now/compose/feature/login/LoginActivity.kt
@@ -2,6 +2,7 @@ package com.sopt.now.compose.feature.login
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
+import androidx.activity.compose.BackHandler
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -57,7 +58,7 @@ fun LoginScreen() {
     val context = LocalContext.current
 
     val loginViewModel =
-        ViewModelProvider(context as ComponentActivity).get(LoginViewModel::class.java)
+        ViewModelProvider(context as ComponentActivity)[LoginViewModel::class.java]
 
     Column(
         modifier = Modifier
